Reorganize project into src/, test/, scripts/ layout
- Move source files to src/
- Move test files to test/ (mirrored structure)
- Move CLI and eval scripts to scripts/
- Remove unused test-ask.ts
- Update all import paths
- Update configs (package.json, jsr.json, docker-compose.yml, justfile, biome.json)
- Update README paths
- Fix type errors
