+++ biome.json
@@ -1,7 +1,7 @@
 {
 	"$schema": "https://biomejs.dev/schemas/2.3.11/schema.json",
 	"files": {
-		"includes": ["**", "!workdir", "!eval/templates", "!eval/report-template.html"]
+		"includes": ["**", "!workdir", "!scripts/eval/report-template.html"]
 	},
 	"assist": { "actions": { "source": { "organizeImports": "on" } } },
 	"linter": {
