+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "ask-forge",
-	"module": "index.ts",
+	"module": "src/index.ts",
 	"type": "module",
 	"scripts": {
 		"format": "biome format --write .",
