+++ src/stream-processor.ts
